--- java/src/main/java/accounting/patterns/Quantity.java
@@ -1,13 +1,0 @@
-package accounting.patterns;
-
-public class Quantity {
-    private final long amount;
-
-    public Quantity(long amount) {
-        this.amount = amount;
-    }
-
-    public long getValue() {
-        return amount;
-    }
-}
