+++ src/main/java/accounting/patterns/Customer.java
@@ -1,32 +1,35 @@
 package accounting.patterns;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Customer {
-    private final ServiceAgreement serviceAgreement;
     private final String name;
-    private List<Entry> entries = new ArrayList<>();
+    private final ServiceAgreement serviceAgreement;
+    private final List<Entry> entries = new ArrayList<>();
 
     public Customer(String name, ServiceAgreement serviceAgreement) {
         this.name = name;
         this.serviceAgreement = serviceAgreement;
     }
 
-    public ServiceAgreement getServiceAgreement() {
-        return serviceAgreement;
-    }
-
-    public void addEntry(Entry newEntry) {
-        entries.add(newEntry);
-
+    public String getName() {
+        return name;
     }
 
     public Entry getEntry(int index) {
         return entries.get(index);
     }
 
-    public String getName() {
-        return name;
+    public void post(EventType eventType, LocalDate eventDate, Quantity quantity) {
+        entries.add(serviceAgreement.post(eventType, eventDate, quantity));
+    }
+
+    public MonetaryAmount balance(LocalDate balanceDate) {
+        return entries.stream()
+                .filter(entry -> entry.getEntryDate().isBefore(balanceDate))
+                .map(Entry::getAmount)
+                .reduce(MonetaryAmount.ZERO, MonetaryAmount::add);
     }
 }
