+++ src/main/java/accounting/patterns/Entry.java
@@ -3,17 +3,21 @@ package accounting.patterns;
 import java.time.LocalDate;
 
 public class Entry {
-    private final LocalDate date;
-    private final EntryType type;
+    private final LocalDate entryDate;
+    private final EntryType entryType;
     private final MonetaryAmount amount;
 
-    public Entry(LocalDate date, EntryType type, MonetaryAmount amount) {
-        this.date = date;
-        this.type = type;
+    public Entry(LocalDate entryDate, EntryType entryType, MonetaryAmount amount) {
+        this.entryDate = entryDate;
+        this.entryType = entryType;
         this.amount = amount;
     }
 
     public MonetaryAmount getAmount() {
         return amount;
+    }
+
+    public LocalDate getEntryDate() {
+        return entryDate;
     }
 }
