+++ src/test/java/accounting/patterns/CustomerTest.java
@@ -1,5 +1,6 @@
 package accounting.patterns;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
@@ -7,23 +8,40 @@ import java.time.LocalDate;
 import static accounting.patterns.EntryType.BASE_USAGE;
 import static accounting.patterns.EventType.USAGE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class CustomerTest {
+
+class CustomerTest {
 
     @Test
-    public void testUsage() {
-        var acm = new Customer("Acme Coffee Makers", new ServiceAgreement(10)
-                .addPostingRule(USAGE, new MultiplyByRatePostingRule(BASE_USAGE), LocalDate.of(1999, 10, 1)));
+    @DisplayName("Posting an event to a customer creates a valid entry")
+    void testPostingAnEvent() {
+        LocalDate effectiveDate = LocalDate.of(1999, 10, 1);
+        Customer customer = new Customer("Acme Coffee Makers",
+                new ServiceAgreement(10)
+                        .addPostingRule(USAGE,
+                                new MultiplyByRatePostingRule(BASE_USAGE),
+                                effectiveDate));
 
-        var usageEvent = new UsageEvent(new Quantity(50), LocalDate.of(1999, 10, 1), acm);
-        usageEvent.process();
+        customer.post(USAGE, effectiveDate, new Quantity(50));
 
-        var resultingEntry = acm.getEntry(0);
+        Entry resultingEntry = customer.getEntry(0);
         assertEquals(new MonetaryAmount("USD", 500), resultingEntry.getAmount());
     }
 
     @Test
-    public void customersAreNamed() throws Exception {
-        assertEquals(new Customer("foo", new ServiceAgreement(1)).getName(), "foo");
+    @DisplayName("Posting an event to a customer without posting rule")
+    void testCustomerWithNoRules() {
+        LocalDate effectiveDate = LocalDate.of(1999, 10, 1);
+        Customer customer = new Customer("Acme Coffee Makers",
+                new ServiceAgreement(10));
+
+        assertThrows(RuntimeException.class, () -> customer.post(USAGE, effectiveDate, new Quantity(50)));
+    }
+
+    @Test
+    @DisplayName("Customers have names")
+    void customersAreNamed() {
+        assertEquals("foo", new Customer("foo", new ServiceAgreement(1)).getName());
     }
 }
