+++ src/test/java/accounting/patterns/EntryTest.java
@@ -8,10 +8,11 @@ import java.util.Currency;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class EntryTest {
+
+class EntryTest {
     @Test
-    public void entriesAreDated() throws Exception {
-        var entry = new Entry(LocalDate.now(), EntryType.entryType("foo"), new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1)));
+    void entriesAreDated() {
+        final Entry entry = new Entry(LocalDate.now(), EntryType.entryType("foo"), new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1)));
 
         assertEquals(new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1)), entry.getAmount());
     }
