+++ src/test/java/accounting/patterns/MonetaryAmountTest.java
@@ -1,5 +1,6 @@
 package accounting.patterns;
 
+
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
@@ -8,11 +9,12 @@ import java.util.Currency;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+
 public class MonetaryAmountTest {
     @Test
     public void differentIfAmountsDiffer() throws Exception {
-        var monetaryAmount1 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
-        var monetaryAmount2 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(2));
+        final MonetaryAmount monetaryAmount1 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
+        final MonetaryAmount monetaryAmount2 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(2));
 
 
         assertFalse(monetaryAmount1.equals(monetaryAmount2));
@@ -20,8 +22,8 @@ public class MonetaryAmountTest {
 
     @Test
     public void differentIfCurrenciesDiffer() throws Exception {
-        var monetaryAmount1 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
-        var monetaryAmount2 = new MonetaryAmount(Currency.getInstance("GBP"), new BigDecimal(1));
+        final MonetaryAmount monetaryAmount1 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
+        final MonetaryAmount monetaryAmount2 = new MonetaryAmount(Currency.getInstance("GBP"), new BigDecimal(1));
 
 
         assertFalse(monetaryAmount1.equals(monetaryAmount2));
@@ -29,8 +31,8 @@ public class MonetaryAmountTest {
 
     @Test
     public void sameIfCurrencyAndValueMatch() throws Exception {
-        var monetaryAmount1 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
-        var monetaryAmount2 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
+        final MonetaryAmount monetaryAmount1 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
+        final MonetaryAmount monetaryAmount2 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
 
 
         assertTrue(monetaryAmount1.equals(monetaryAmount2));
@@ -38,10 +40,10 @@ public class MonetaryAmountTest {
 
     @Test
     public void hashcodeVariesByCurrencyAndValue() throws Exception {
-        var monetaryAmount1 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
-        var monetaryAmount2 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
-        var monetaryAmount3 = new MonetaryAmount(Currency.getInstance("GBP"), new BigDecimal(1));
-        var monetaryAmount4 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(2));
+        final MonetaryAmount monetaryAmount1 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
+        final MonetaryAmount monetaryAmount2 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(1));
+        final MonetaryAmount monetaryAmount3 = new MonetaryAmount(Currency.getInstance("GBP"), new BigDecimal(1));
+        final MonetaryAmount monetaryAmount4 = new MonetaryAmount(Currency.getInstance("USD"), new BigDecimal(2));
 
 
         assertTrue(monetaryAmount1.hashCode() == monetaryAmount2.hashCode());
